Added a new `clipping` submodule to `layout` and moved the clipping systems there. Also added a basic regression test.
